Changed the grid finding method to make it less error prone.
Adjusted the way the program finds the edges of the item. Previously it was prone to errors because of false edges inside the item. Adjusted the way it searches for items, to start with a small bounding box that grows as it finds new edges until it finds an acceptable set of edges. This on average takes about 2x the CPU time with the top end limited because of the smaller search area. However, worst case can lead to 10x the CPU time. Prior to the change in these worst case scenarios, the scan failed more often than not.

It is not perfect. Currently when viewing the stash, items towards the top center of the screen will have difficulty scanning. This is due to the wash out effect from the 'ceiling light' on the menu. Previously it also failed in this case. It can also sometimes fail if the game tooltip is up and positioned just right. It scales the edge score based on distance from center to try to correct for this. However, it isn't perfect. I think this is acceptable because the image classifier also frequently fails when the tooltip is up, and the previous edge detector also would sometimes fail when the tooltip was up. So ideally you scan prior to the tooltip popping up.