Adjust grid finding algorithm in item scan to be more accurate. (#91)
Adjusted the way the program finds the edges of the item. Previously it was prone to errors because of false edges inside the item. Adjusted the way it searches for items, to start with a small bounding box that grows as it finds new edges until it finds an acceptable set of edges. This on average takes about 2x the CPU time with the top end limited because of the smaller search area. However, worst case can lead to 10x the CPU time. Prior to the change in these worst case scenarios, the scan failed more often than not.